# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StockSense inventory management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Error"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Error"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Current session user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Generate a development password reset OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: OTP generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
        "404":
          $ref: "#/components/responses/Error"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset a password with an OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [USER, ADMIN]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
    SignupInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordInput:
      type: object
      required: [email, otp, password]
      properties:
        email: { type: string, format: email }
        otp: { type: string, minLength: 6, maxLength: 6 }
        password: { type: string, minLength: 8 }
    OtpResponse:
      type: object
      required: [message, otp]
      properties:
        message: { type: string }
        otp: { type: string }
...