# lib/api-zod/src/generated/types/otpResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * StockSense inventory management API
 * OpenAPI spec version: 0.1.0
 */

export interface OtpResponse {
  message: string;
  otp: string;
}